feat: add support for drag and drop directories or multiple files
